Add imports for both kinds of taskflows
This adds the import for the container pull taskflow
as well.
